feat: 수동 로또 입력 로직 추가 (미완)
