+++ src/main/java/app/strategy/LottoNumStrategy.java
@@ -1,0 +1,7 @@
+package app.strategy;
+
+import app.domain.lotto.Lotto;
+
+public interface LottoNumStrategy {
+    Lotto makeLotto();
+}
