--- src/main/java/lotto/action/LottoMachine.java
@@ -1,29 +1,0 @@
-package lotto.action;
-
-import lotto.domain.Lotto;
-import lotto.strategy.LottoNumStrategy;
-
-public class LottoMachine {
-
-    private LottoNumStrategy strategy;
-
-    private LottoMachine() {
-    }
-
-    private LottoMachine(LottoNumStrategy strategy) {
-        this.strategy = strategy;
-    }
-
-    public static LottoMachine create() {
-        return new LottoMachine();
-    }
-
-    public static LottoMachine of(LottoNumStrategy strategy) {
-        return new LottoMachine(strategy);
-    }
-
-    public Lotto makeLotto() {
-        return strategy.makeLotto();
-    }
-
-}
