--- src/main/java/lotto/domain/Money.java
@@ -1,15 +1,0 @@
-package lotto.domain;
-
-public class Money {
-    private int money;
-
-    private Money(int money) {
-        validation(money);
-        this.money = money;
-    }
-
-    private void validation(int money) {
-    }
-
-
-}
