--- src/main/java/lotto/strategy/LottoNumStrategy.java
@@ -1,7 +1,0 @@
-package lotto.strategy;
-
-import lotto.domain.Lotto;
-
-public interface LottoNumStrategy {
-    Lotto makeLotto();
-}
