--- src/test/java/lotto/domain/MoneyTest.java
@@ -1,16 +1,0 @@
-package lotto.domain;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
-
-class MoneyTest {
-
-    @DisplayName("돈 객체는 양수만 가능하다.")
-    @Test
-    void IsOnlyPositive(){
-
-    }
-}
